+++ package.json
@@ -2,10 +2,11 @@
     "name": "gb_js_2",
     "version": "1.0.0",
     "description": "JS2 study",
-    "main": "index.js",
+    "main": "server.js",
     "scripts": {
         "test": "echo \"Error: no test specified\" && exit 1",
-        "start": "node index.js"
+        "start": "node server.js",
+        "build": "webpack"
     },
     "repository": {
         "type": "git",
@@ -20,5 +21,15 @@
     "bugs": {
         "url": "https://github.com/CuranderoRu/GB_JS_2/issues"
     },
-    "homepage": "https://github.com/CuranderoRu/GB_JS_2#readme"
+    "homepage": "https://github.com/CuranderoRu/GB_JS_2#readme",
+    "devDependencies": {
+        "webpack": "^5.6.0",
+        "webpack-cli": "^4.2.0"
+    },
+    "dependencies": {
+        "@babel/core": "^7.12.8",
+        "@babel/plugin-proposal-class-properties": "^7.12.1",
+        "babel": "^6.23.0",
+        "babel-loader": "^8.2.1"
+    }
 }
