+++ public/index.html
@@ -7,11 +7,6 @@
     <title>JS level 2 study environment</title>
     <link rel="stylesheet" href="./css/style.css">
     <script src="./js/jquery-3.5.1.min.js"></script>
-    <script src="./js/app.js"></script>
-    <script src="./js/main.js"></script>
-    <script>
-        window.onload = init;
-    </script>
 </head>
 
 <body>
@@ -67,18 +62,6 @@
                 <!-- <div class="cart-item"><span>ESP8266 NodeMCU v3</span><a class="cart-item-button" href="">-</a><span>1</span> <a class="cart-item-button" href="">+</a> <span> x </span><span>299.97</span><span> = </span><span>299.97 ₽</span><a class="cart-item-button" href="">X</a></div> -->
             </section>
             <section class="main-support">
-                <!-- <form onsubmit="">
-                    <span>Contact us</span>
-                    <label for="input_name">Your name
-                    <input id="input_name" type="text" placeholder ="Ivan" required></label>
-                    <label for="input_phone">Your phone number
-                    <input id="input_phone" type="text" placeholder ="+7(000)000-0000"></label>
-                    <label for="input_email">Your contact e-mail
-                    <input id="input_email" type="text" placeholder ="my@mail.com" required></label>
-                    <label for="input_content">Your message:</label>
-                    <textarea name="" id="input_content" cols="50" rows="10"></textarea>
-                    <input id="input_submit" type="submit" value="Send">
-                </form> -->
             </section>
         </main>
         <footer class="footer">
@@ -94,6 +77,7 @@
             </section>
         </footer>
     </div>
+    <script src="./js/shop.js"></script>
     <script>
         $('#menu__toggle').on('change', function() {
             if ($(this).prop('checked')) {
