+++ public/js/shop.js
@@ -1,0 +1,1 @@
+(()=>{"use strict";class t{constructor(t,e){this._id=t,this.className=e}render(){let t=document.createElement("div");return t.id=this.id,t.classList.add(this.className),t}get id(){return this._id}}class e extends t{constructor({id:t,title:e,currency:s,price:r,category:a,imgsrc:n,popularity:i},l="",o="product-item",c){super(t,o),this.imgsrc=l||n,this.title=e,this.currency=s,this.price=r,this.category=a,this.popularity=i,this.cart=c}render(){let t=document.createElement("div");t.id=this.id,t.classList.add(this.className);let e=document.createElement("img");e.src=this.imgsrc,e.width="250",e.height="250",t.append(e);let s=document.createElement("div");s.classList.add("product-item-spec");let r=document.createElement("h3");r.textContent=this.title,s.append(r);let a=document.createElement("span");a.classList.add("product-item-spec-price"),a.innerHTML=`${this.currency} ${this.price}`,s.append(a);let n=document.createElement("a");return n.href="#",n.id=`buybutton-${this.id}`,n.classList.add("product-item-spec-button"),n.textContent="В корзину",n.addEventListener("click",this.cart),s.append(n),t.append(s),t}}class s extends t{constructor(t,e,s){super(t,"main-section-btn"),this.textContent=e,this.pushHandler=s}render(){let t=document.createElement("button");return t.classList.add(this.className),t.textContent=this.textContent,t.id=this.id,t.addEventListener("click",this.pushHandler),t}}class r{hideContent(){let t=document.getElementsByClassName("main-section");for(let e=0;e<t.length;e++)t[e].classList.add("invisible");t=document.getElementsByClassName("cart");for(let e=0;e<t.length;e++)t[e].classList.add("invisible");t=document.getElementsByClassName("main-support");for(let e=0;e<t.length;e++)t[e].classList.add("invisible")}showProducts(){let t=document.getElementsByClassName("main-section");for(let e=0;e<t.length;e++)t[e].classList.remove("invisible")}showCart(){let t=document.getElementsByClassName("cart")[0];return t.classList.remove("invisible"),t}showSupport(){let t=document.getElementsByClassName("main-support")[0];return t.classList.remove("invisible"),t}}class a{constructor(t=""){this.path=t.replace(document.location.origin+"/","")}}class n{getMenuArr(){return[{href:"controllers",label:"Controllers",submenu:[{href:"controllers/esp",label:"ESP",category:"ESP"},{href:"controllers/arduino",label:"Arduino",category:"Arduino"},{href:"controllers/raspberry",label:"Raspberry",category:"Raspberry"},{href:"controllers/stm",label:"STM",category:"STM"}]},{href:"periferals",label:"Periferals",submenu:[{href:"periferals/thermosensors",label:"Thermosensors"},{href:"periferals/air-quality",label:"Air quality"},{href:"periferals/relay",label:"Relay"}]},{href:"#",label:"About"},{href:"#",label:"Blog"},{href:"support",label:"Support"}]}buildMenu(t){let e=this.getMenuArr();return e.forEach(((e,s)=>{let r;if("submenu"in e&&e.submenu instanceof Array&&e.submenu.length>0){let a=e.submenu.map((e=>({item:new l("menuCategory"+e.category,e.href,"page-header-nav-item-submenu-item","page-header-nav-item-link",e.label,void 0,t)})));r=new i("nav-menu-submenu-"+s,"page-header-nav-item-submenu",a)}e.item=new l(null,e.href,"page-header-nav-item","page-header-nav-item-link",e.label,r,t)})),new i("nav-menu","page-header-nav-menu",e)}buildContactForm(){return new o}buildCartMenu(t){if(!document.getElementById("cartManagementMenu")){let e=document.getElementsByClassName("page-header-cart")[0],s=new i("cartManagementMenu","page-header-cart-actions",[{item:new l("viewCartItems","#",null,"page-header-nav-item-link","View items",null,t)},{item:new l("clearCart","#",null,"page-header-nav-item-link","Clear cart",null,t)}]);e.append(s.render())}}}class i extends t{constructor(t,e,s){super(t,e),this.items=s}render(){let t=document.createElement("ul");t.id=this.id,t.classList.add(this.className);for(let e=0;e<this.items.length;e++)this.items[e].item instanceof l&&t.append(this.items[e].item.render());return t}}class l extends t{constructor(t,e,s,r,a,n,i){super(t,s),this.href=e,this.aclass=r,this.label=a,this.submenu=n,this.handler=i}render(){let t=document.createElement("li");this.className&&t.classList.add(this.className);let e=document.createElement("a");return e.href=this.href,this.aclass&&e.classList.add(this.aclass),this.id&&(t.id=this.id,e.id=`a${this.id}`),e.textContent=this.label,this.handler&&e.addEventListener("click",this.handler),t.append(e),this.submenu&&t.append(this.submenu.render()),t}}class o{check(t,e){return e.test(t)}handleEvent(t){t.preventDefault();let e=document.getElementById("input_name");this.check(e.value,/^[A-Za-zА-Яа-яЁё\ ]+$/)?e.classList.remove("incorrect_input"):(e.classList.add("incorrect_input"),console.log(e.value));let s=document.getElementById("input_phone");this.check(s.value,/\+[\d]{1}\([\d]{2,3}\)[\d]{2,3}-[\d]{2,3}-[\d]{2,3}$/)?s.classList.remove("incorrect_input"):(console.log(s.value),s.classList.add("incorrect_input"));let r=document.getElementById("input_email");this.check(r.value,/^[\w]{1}[\w-\.]*@[\w-\.]+\.[a-z]{2,4}$/i)?r.classList.remove("incorrect_input"):(r.classList.add("incorrect_input"),console.log(r.value))}render(){let t=document.createElement("form");t.addEventListener("submit",this),t.classList.add("contact-form");let e=document.createElement("span");e.textContent="Contact us",t.append(e);let s=document.createElement("label");s.for="input_name",s.textContent="Your name";let r=document.createElement("input");r.id="input_name",r.type="text",r.placeholder="Ivan",r.required=!0,s.append(r),t.append(s);let a=document.createElement("label");a.for="input_phone",a.textContent="Your phone number";let n=document.createElement("input");n.id="input_phone",n.type="text",n.placeholder="+7(000)000-0000",a.append(n),t.append(a);let i=document.createElement("label");i.for="input_email",i.textContent="Your contact e-mail";let l=document.createElement("input");l.id="input_email",l.type="text",l.placeholder="my@mail.com",i.append(l),t.append(i);let o=document.createElement("label");o.for="input_content",o.textContent="Your message:",t.append(o);let c=document.createElement("textarea");c.id="input_content",c.cols="50",c.rows="10",t.append(c);let d=document.createElement("input");return d.id="input_submit",d.type="submit",d.value="Send",t.append(d),t}}class c{constructor(t=[]){this._version="1.0",this._cartSum=0,this._items=[],this.className="cart",this._productsArray=t,this.init()}get version(){return this._version}get productsArray(){return this._productsArray}set productsArray(t){this._productsArray=t}get items(){return this._items}set items(t){!t instanceof Array||(this._items=[],t.forEach((function(t,e,s){this._items.push(new d(t,t._q,this))})))}loadItems(t){!t instanceof Array||(this._items=[],t.forEach((function(t,e,s){this._items.push(new d(t,t._q,this))})))}init(){(new n).buildCartMenu(this),this.recalcCostLabel()}add(t,e=1){return new Promise(((s,r)=>{let a=!1,n=this._items.find((e=>e.id===t));if(void 0!==n&&(n.q=n.q+e,n.render(!0),a=!0),!a){let e=this.getProductById(t);void 0!==e&&(this._items.push(new d(e,1,this)),a=!0,localStorage.setItem("cart",JSON.stringify(this.getStorageData())))}this.total(),this.init(),a?s():r()}))}getStorageData(){let t=this._items.map((t=>({id:t.id,_q:t._q})));return{version:this._version,_items:t}}remove(t,e=0){return new Promise(((s,r)=>{let a=!1,n=this._items.findIndex((e=>e.id===t));n>=0&&(0===e||e>=this._items[n].q?(this._items.splice(n,1),this.render()):(this._items[n].q=this._items[n].q-e,this._items[n].render(!0)),a=!0),this.total(),a?s():r()}))}clear(){this._items=[],this._cartSum=0,document.querySelector(".cart").innerHTML="",localStorage.setItem("cart",JSON.stringify(this.getStorageData()))}total(){this._cartSum=0;for(let t=0;t<this._items.length;t++)this._cartSum+=this._items[t].sum;this._cartSum=this._cartSum.toFixed(2)}get cartSum(){return this._cartSum}getProductById(t){return this._productsArray.find((e=>e.id===t))}recalcCostLabel(){document.getElementById("cart-cost").innerHTML="&#8381; "+this._cartSum}handleEvent(t){if(t.preventDefault(),"aclearCart"===t.target.id)return this.clear(),document.getElementById("cartManagementMenu").remove(),void this.recalcCostLabel();if("aviewCartItems"===t.target.id){let t=document.getElementsByClassName("main-section");for(let e=0;e<t.length;e++)t[e].classList.toggle("invisible");return document.querySelector(".cart").classList.toggle("invisible"),void this.render()}let e=t.target.id.split("-");switch(e[0]){case"buybutton":this.add(e[1]).then((()=>{this.recalcCostLabel()})).catch((()=>{console.log("Could not add good to cart")}));break;case"remove":this.remove(e[1],1).then((()=>{this.recalcCostLabel()})).catch((()=>{console.log("Could not remove good from cart")}));break;case"dropbutton":this.remove(e[1]).then((()=>{this.recalcCostLabel()})).catch((()=>{console.log("Could not remove good from cart")}));break;default:console.log("Unsupported call ",e[0])}}render(){let t=document.getElementsByClassName(this.className)[0];t.innerHTML="",this._items.forEach((function(e,s,r){t.append(e.render())}))}}class d extends e{constructor(t,e,s){super(t,"","cart-item",s),this._q=e,this._sum=+(e*t.price).toFixed(2)}get sum(){return this._sum}get q(){return this._q}set q(t){this._q=t,this._sum=+(t*this.price).toFixed(2)}render(t=!1){let e=`cartProduct-${this.id}`,s=document.getElementById(e);if(s)s.innerHTML="";else{if(t)return;s=document.createElement("div"),s.id=e,s.classList.add(this.className)}let r=document.createElement("span");r.textContent=this.title;let a=document.createElement("a");a.textContent="-",a.className="cart-item-button",a.id=`remove-${this.id}`,a.addEventListener("click",this.cart);let n=document.createElement("a");n.textContent="+",n.className="cart-item-button",n.id=`buybutton-${this.id}`,n.addEventListener("click",this.cart);let i=document.createElement("a");i.textContent="X",i.className="cart-item-button",i.id=`dropbutton-${this.id}`,i.addEventListener("click",this.cart);let l=document.createElement("span");l.textContent=this._q;let o=document.createElement("span");o.textContent="x";let c=document.createElement("span");c.textContent=this.price;let d=document.createElement("span");d.textContent="=";let u=document.createElement("span");return u.textContent=`${this.sum} ${this.currency}`,s.append(r),s.append(a),s.append(l),s.append(n),s.append(o),s.append(c),s.append(d),s.append(u),s.append(i),s}}function u(t,e,s){return e in t?Object.defineProperty(t,e,{value:s,enumerable:!0,configurable:!0,writable:!0}):t[e]=s,t}new class{constructor(t){if(u(this,"name","app"),u(this,"productsArray",[]),this.cart=new c,void 0===localStorage.cart)this.initLocalStorage();else{let t=JSON.parse(localStorage.cart);this.cart.version!==t.version&&this.initLocalStorage()}this.smallImagePath=t.smallImagePath,this.bigImagePath=t.bigImagePath,this.pageIndex=0,this.prod_start_pos=0,this.route=new a,this.fetchProducts().then((t=>{this.productsArray=t.map((t=>new e(t,this.smallImagePath+t.imgsrc,"product-item",this.cart))),this.cart.productsArray=this.productsArray,this.fillProducts("latestList"),this.prod_start_pos=this.prod_start_pos+3,this.fillProducts("popularList")})).catch((()=>{this.cart.productsArray=[]}))}initLocalStorage(){localStorage.setItem("cart",JSON.stringify(this.cart.getStorageData()))}getPriceById(t){return this.productsArray.filter((e=>e.id===t))[0].price}fillProducts(t,e=null){let r,a=document.getElementById(t);if("latestList"===t){let t=document.getElementById("mainListHeader");e?(r=this.getProductsByCategory(e),t.textContent=e,a.innerHTML=""):(t.textContent="Latest Products",r=this.getLatestProductsArray(this.prod_start_pos,this.prod_start_pos+3))}else r=this.getPopularProductsArray();for(let t=0;t<r.length;t++)a.append(r[t].render());if("latestList"===t){let t=document.getElementById("loadMoreBTN");t||(t=new s("loadMoreBTN","Load more",this),a.parentNode.append(t.render()))}}getLatestProductsArray(t=0,e=3){return this.productsArray.filter(((s,r)=>r>=t&&r<e))}getPopularProductsArray(){let t=[],e=[].concat([],this.productsArray);return e.sort(((t,e)=>e.popularity-t.popularity)),e.reduce(((t,e)=>(t.length<=2&&e.popularity>0&&t.push(e),t)),t),t}getProductsByCategory(t){return this.productsArray.filter((e=>e.category.indexOf(t)>=0))}getProductById(t){return this.productsArray.find((e=>e.id===t))}fetchProducts(){return fetch(`/js/database${this.pageIndex}.json`).then((t=>t.json())).then((t=>t.data)).catch((t=>{console.warn("Check your network connection",t)}))}handleEvent(t){if(t.preventDefault(),"loadMoreBTN"===t.target.id)this.prod_start_pos+3>this.productsArray.length?(this.pageIndex=this.pageIndex+1,this.fetchProducts().then((t=>{this.productsArray=this.productsArray.concat(t.map((t=>new e(t,this.smallImagePath+t.imgsrc,"product-item",this.cart)))),this.cart.productsArray=this.productsArray,this.fillProducts("latestList",this.route.path),this.prod_start_pos=this.prod_start_pos+3})).catch((()=>{this.pageIndex=this.pageIndex-1}))):(this.fillProducts("latestList",this.route.path),this.prod_start_pos=this.prod_start_pos+3);else{this.route=new a(t.target.href);let e=new r;switch(e.hideContent(),this.route.path){case"support":e.showSupport();break;default:e.showProducts(),this.fillProducts("latestList",this.route.path)}}}buildForm(){const t=document.getElementById("nav"),e=new n;t.append(e.buildMenu(this).render()),document.getElementsByClassName("main-support")[0].append(e.buildContactForm().render());const s=new r;s.hideContent(),s.showProducts()}}({smallImagePath:"./img/small/",bigImagePath:"./img/big/"}).buildForm()})();
