+++ src/route.js
@@ -1,0 +1,7 @@
+'use strict';
+
+export default class Route {
+    constructor(href = '') {
+        this.path = href.replace(document.location.origin + '/', '');
+    }
+}
